Fix local installation
`pip install -e .` was failing.
The previous setup was creating a libmagent.so file, but the build
script was looking for a cpython specific library file. This failed
because it didn't exist, breaking the build. It then tried to load
the libmagent.so file which wasn't correct because it was supposed
to use the cpython specific file.

Fixing both of these issues, it now installs with pip install -e .
This works on linux, probably works on mac. I have no idea on windows

I do not know if this will package correctly for pip install with pypi